do this only if installed (allows split of tools/utils)

+++ tools/bootstrap.sh
@@ -5,8 +5,11 @@ if [ ! -d "$PWD/users" ]; then
 fi
 DIR=/usr/share/radiucal/
 for s in $(echo "configure monitor reports"); do
-    cp $DIR/$s.sh $s
-    chmod u+x $s
+    p=$DIR/$s.sh
+    if [ -e $p ]; then
+        cp $DIR/$s.sh $s
+        chmod u+x $s
+    fi
 done
 for f in $(echo "netconf accounts users/__config__ users/__init__"); do
     cp $DIR/$f.py $PWD/$f.py
